* playback-transport:
  UI: b/positionview.vue: adjust to sixteenth telemetry
  ASE: project.cc: adjust Transport.running and adapt current_sixteenth telemetry
  UI: b/partlist.vue: disable unfinished logic
  UI: b/editable.vue: use #000 as input bg and auto-convert "clicks"
  UI: util.js: remove unused v-inlineblur
  UI: b/trackview.vue: use b-editable to edit track labels
  UI: b/contextmenu.vue: prevent spurious enter/leave/etc events during transitions
  UI: b/editable.vue: allow explicit and double click activation
  UI: b/positionview.vue: allow BPM changes via b-editable
  UI: b/positionview.vue: remove unused code
  UI: b/editable.vue: display an editable <span/>
  ASE: jsonapi.cc: require 'Origin: http://localhost' for socket connections
  ASE: websocket: provide header() getter for WebSocketConnection::Info
  ASE: project: implement real BPM setter
  ASE: properties: support Range() property with explicit getter and setter
  ASE: api.hh: Gadget: implement list_properties(), get_value(), set_value()
  UI: b/positionview.vue: add input field
  UI: b/positionview.vue: move frequent updates into text node
  UI: b/trackview.vue: avoid @contextmenu propagation out of INPUT field
  UI: startup.js: allow button3 browser menus in INPUT fields
  UI: b/positionview.vue: display bar.beat.sixteenth.fraction minutes:seconds
  UI: util.js: support Int8 telemtry data
  ASE: project.cc: export current bar/beat/sixteenth, minutes/seconds, etc
  ASE: project: add autoplay timer
  ASE: server.hh: add int8 telemetry_type
  ASE: cxxaux.hh: add divmod() helper
  ASE: loop: add clear_source() helper
  UI: util.js: export extend_property()
  UI: b/positionview.vue: display bpm, numerator, denominator
  ASE: project: add musical_tuning, dirty, bpm, numerator, denominator properties
  ASE: server: provide EnumInfo::impl for MusicalTuning
  ASE: value: add EnumInfo to augment enumerations
  ASE: properties: add ident, support float Range and extend Enum properties
  JSONIPC: jsonipc.hh: add Enum<>::list_values()
  ASE: gadget.cc: implement access_property() in terms of access_properties()
  ASE: clip.cc: notify about note changes
  UI: b/positionview.vue: show current_bpm and current_tick
  ASE: project.cc: provide current_bpm and current_tick as double
  UI: util.js: remove unused telemetry type "u32"
  ASE: server.hh: remove unused telemetry type "u32"
  ASE: api.hh: use 64bit for ticks and duration
  ASE: engine: leave timing defaults to AudioTransport ctor
  ASE: engine: use transport->advance()
  UI: b/*.vue: use Util.PPQN
  UI: util.js: export PPQN as 4838400
  ASE: clip.cc: use TRANSPORT_PPQN
  ASE: transport: provide TickSignature to handle all AudioTransport timing
	Squashed commits to introduce TickSignature:
	  Move initialisation into AudioTransport()
	  Use 64bit for ticks and duration
	  Quantize current_tick_d
	  Provide current bar/beat/sixteenth, minutes/seconds
	  Add next_bar_tick()
	  Add TimeSignature class
	  Add set_tick(), tick_from_beat(), tick_from_time()
	  Add TimeSignature tests
	  Change PPQN to 4838400
	  Add TickSignature for musical time signature and tick conversions
	  Support getters and samplerate in TickSignature
	  Add sample_to_tick()
	  Add TickSignature::sample_from_tick()
	  Provide current_bar_tick, sample_to_tick(), sample_from_tick()
	  Add const to TickSignature methods
	  Add beats_per_bar() and beat_unit()
	  Constrain set_signature() inputs and return changed flag
	  Remove unused numerator and denominator
	  Construct TickSignature with defaults to avoid 0-division

+++ ase/clip.hh
@@ -13,7 +13,7 @@ constexpr const uint MIDI_NOTE_ID_FIRST = 0x10000001;
 constexpr const uint MIDI_NOTE_ID_LAST = 0xfffffffe;
 
 class ClipImpl : public GadgetImpl, public virtual Clip {
-  int32 starttick_ = 0, stoptick_ = 0, endtick_ = 0;
+  int64 starttick_ = 0, stoptick_ = 0, endtick_ = 0;
   struct CmpNoteTicks { int operator() (const ClipNote &a, const ClipNote &b) const; };
   struct CmpNoteIds   { int operator() (const ClipNote &a, const ClipNote &b) const; };
   EventList<ClipNote,CmpNoteIds> notes_;
@@ -29,13 +29,13 @@ protected:
 public:
   using OrderedEventsP = OrderedEventsV::ConstP;
   OrderedEventsP tick_events    ();
-  int32          start_tick     () const override { return starttick_; }
-  int32          stop_tick      () const override { return stoptick_; }
-  int32          end_tick       () const override { return endtick_; }
-  void           assign_range   (int32 starttick, int32 stoptick) override;
+  int64          start_tick     () const override { return starttick_; }
+  int64          stop_tick      () const override { return stoptick_; }
+  int64          end_tick       () const override { return endtick_; }
+  void           assign_range   (int64 starttick, int64 stoptick) override;
   ClipNoteS      list_all_notes () override;
   bool           needs_serialize() const;
-  int32          change_note    (int32 id, int32 tick, int32 duration, int32 key, int32 fine_tune, double velocity) override;
+  int32          change_note    (int32 id, int64 tick, int64 duration, int32 key, int32 fine_tune, double velocity) override;
   ASE_DEFINE_MAKE_SHARED (ClipImpl);
 };
